Move createCartFromLineItems to own file
