+++ packages/composite-checkout/src/lib/payment-methods/paypal.js
@@ -12,6 +12,7 @@ import { useLocalize } from '../../lib/localize';
 import { useDispatch, useSelect, usePaymentData } from '../../lib/registry';
 import { useCheckoutHandlers, useCheckoutRedirects, useLineItems } from '../../public-api';
 import { PaymentMethodLogos } from '../styled-components/payment-method-logos';
+import { createCartFromLineItems } from './transactions-endpoint';
 
 export function createPayPalMethod( { registerStore, makePayPalExpressRequest } ) {
 	registerStore( 'paypal', {
