+++ packages/composite-checkout/src/lib/payment-methods/stripe-credit-card-fields.js
@@ -42,6 +42,7 @@ import { SummaryLine, SummaryDetails } from '../styled-components/summary-detail
 import CreditCardFields, { CVVImage } from './credit-card-fields';
 import Spinner from '../../components/spinner';
 import ErrorMessage from '../../components/error-message';
+import { createCartFromLineItems } from './transactions-endpoint';
 
 export function createStripeMethod( {
 	getSiteId,
@@ -773,36 +774,5 @@ function formatDataForTransactionsEndpoint( {
 		} ),
 		domainDetails,
 		payment,
-	};
-}
-
-function createCartFromLineItems( {
-	siteId,
-	couponId,
-	items,
-	country,
-	postalCode,
-	subdivisionCode,
-} ) {
-	const currency = items.reduce( ( firstValue, item ) => firstValue || item.amount.currency, null );
-	return {
-		blog_id: siteId,
-		coupon: couponId || '',
-		currency: currency || '',
-		temporary: false,
-		extra: [],
-		products: items.map( item => ( {
-			product_id: item.id,
-			meta: '', // TODO: get this for domains, etc
-			currency: item.amount.currency,
-			volume: 1,
-		} ) ),
-		tax: {
-			location: {
-				country_code: country,
-				postal_code: postalCode,
-				subdivision_code: subdivisionCode,
-			},
-		},
 	};
 }
